+++ src/plugins/graphql/graphqlProtocolReplacer.spec.ts
@@ -3,12 +3,13 @@ import { graphqlProtocolReplacer } from './graphqlProtocolReplacer';
 
 describe('graphqlProtocolReplacer', () => {
   it('should return POST', async () => {
-    const request: models.Request = { method: 'graphql' };
+    const request: models.Request = { url: '', method: 'graphql' };
     await graphqlProtocolReplacer(request);
     expect(request.method).toEqual('POST');
+    expect(request.headers?.['content-type']).toEqual('application/json');
   });
   it('should  not change', async () => {
-    const request: models.Request = { method: 'get' };
+    const request: models.Request = { url: '', method: 'get' };
     await graphqlProtocolReplacer(request);
     expect(request.method).toEqual('get');
   });